+++ c_test_environment/clang_tests.py
@@ -16,7 +16,7 @@ import os
 class DatalogClangTest(unittest.TestCase, DatalogPlatformTest):
     def check(self, query, name):
         with Chdir("c_test_environment") as d:
-            os.remove("%s.cpp" % name)
+            os.remove("%s.cpp" % name) if os.path.exists("%s.cpp" % name) else None
             emitCode(query, name, CCAlgebra)
             checkquery(name, ClangRunner())
